requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==23.2.1
aiogram==2.25.1
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
antlr4-python3-runtime==4.9.2
anyio==3.7.1
APScheduler==3.9.1.post1
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
autopep8==1.5.6
aws-wsgi==0.2.7
Babel==2.9.1
beautifulsoup4==4.12.3
bidict==0.22.1
boto3==1.28.48
botocore==1.31.85
CacheControl==0.12.14
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==7.1.2
cloudpickle==2.2.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
comm==0.2.1
confluent-kafka==2.2.0
cryptography==41.0.3
cx-Oracle==8.3.0
cycler==0.10.0
dask==2022.10.2
dataclasses-json==0.6.0
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.6
Django==3.2.23
django-cors-headers==4.1.0
djangorestframework==3.14.0
dnspython==2.6.0
docarray==0.39.1
docopt==0.6.2
docutils==0.16
exceptiongroup==1.1.3
executing==2.0.1
fbmessenger==6.0.0
filelock==3.8.2
fire==0.5.0
Flask==1.1.2
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.0
future==0.18.3
gast==0.4.0
google-pasta==0.2.0
graphviz==0.20.1
greenlet==1.0.0
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
ifaddr==0.2.0
importlib-metadata==6.8.0
install-jdk==0.3.0
ipykernel==6.29.0
ipython==8.18.1
itsdangerous==1.1.0
jax==0.4.14
JayDeBeApi==1.2.3
jedi==0.19.1
Jinja2==2.11.3
jmespath==1.0.1
joblib==1.2.0
JPype1==1.4.1
jsonpatch==1.33
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.17.3
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.12.0
kiwisolver==1.3.1
langchain==0.0.310
langchain-core==0.0.13
langchain-experimental==0.0.27
langsmith==0.0.69
libclang==16.0.6
locket==1.0.0
magic-filter==1.0.11
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==1.1.1
marshmallow==3.20.1
matplotlib==3.4.1
matplotlib-inline==0.1.6
mattermostwrapper==2.2
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.0.5
multidict==5.2.0
mypy-extensions==1.0.0
mysqlclient==2.0.3
nest-asyncio==1.6.0
networkx==2.6.3
numexpr==2.8.6
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
orjson==3.9.10
packaging==23.2
pamqp==3.2.1
pandas==1.2.3
parso==0.8.3
partd==1.4.0
peewee==3.14.4
pexpect==4.9.0
Pillow==8.2.0
pipreqs==0.4.13
platformdirs==2.6.0
plotly==4.14.3
pluggy==1.3.0
portalocker==2.7.0
prompt-toolkit==3.0.43
protobuf==4.23.3
psutil==5.9.8
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py_eureka_client==0.11.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.7.0
pycparser==2.21
pydantic==1.10.9
pydot==1.4.2
Pygments==2.17.2
PyJWT==2.8.0
pykwalify==1.8.0
pymongo==4.6.0
pyparsing==2.4.7
pypdf==3.17.4
pyreadline3==3.4.1
pyrsistent==0.19.3
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==0.21.1
python-editor==1.0.4
python-engineio==4.7.1
python-socketio==5.9.0
pytz==2021.1
PyYAML==6.0.1
pyzmq==25.1.2
questionary==1.10.0
randomname==0.1.5
rank-bm25==0.2.2
redis==4.6.0
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.3
rich==13.7.0
rocketchat-API==1.30.0
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.2
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.11.2
sentry-sdk==1.40.4
simple-websocket==0.10.1
six==1.15.0
sklearn-crfsuite==0.3.6
slack-sdk==3.22.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.44
sqlparse==0.4.4
stack-data==0.6.3
structlog==23.1.0
structlog-sentry==2.0.3
tabulate==0.9.0
tarsafe==0.0.4
tenacity==8.2.3
tensorboard-data-server==0.7.1
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.2.0
tiktoken==0.5.1
todo==0.1
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
twilio==8.2.2
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing==3.7.4.3
typing-inspect==0.9.0
typing-utils==0.1.0
typing_extensions==4.7.1
tzdata==2022.7
tzlocal==5.0.1
ujson==5.8.0
urllib3==1.26.18
virtualenv==20.17.1
wcwidth==0.2.6
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==1.0.1
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.0.1
yarg==0.1.9
yarl==1.9.2
zipp==3.16.2
